% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_openai_init}
\alias{azure_openai_init}
\title{Setup environment for Azure use and token tracking}
\usage{
azure_openai_init()
}
\value{
Message
}
\description{
When the session ends, token usage info will be uploaded to a Google Drive
}
