% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_openai_setup}
\alias{azure_openai_setup}
\title{Setting up the Azure API info and storing it in the keyring}
\usage{
azure_openai_setup(username, force_overwrite = F, key)
}
\arguments{
\item{username}{Username can be anything. It's used to distinguish between
users who share an API key}

\item{force_overwrite}{(Default  = FALSE) Force reset of the parameters}

\item{key}{(Optional) Azure API key. Pop-up will show if not provided}
}
\value{
Message if success
}
\description{
Setting up the Azure API info and storing it in the keyring
}
