% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{cred_app}
\alias{cred_app}
\title{Interactive app to set Azure credentials}
\usage{
cred_app()
}
\value{
Starts a Shiny app that returns username and key
}
\description{
Interactive app to set Azure credentials
}
