% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{cred_check}
\alias{cred_check}
\title{Validate Azure API key}
\usage{
cred_check(key, error_on_fail = F)
}
\arguments{
\item{key}{API / Token to check}

\item{error_on_fail}{(Default = F) Generate error when failed authentication}
}
\value{
List with checks
}
\description{
Validate Azure API key
}
