% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{edAI_setup}
\alias{edAI_setup}
\title{Setting up the Azure API info and storing it in the keyring}
\usage{
edAI_setup(force_overwrite = F, username, key)
}
\arguments{
\item{force_overwrite}{(Default  = FALSE) Force reset of the parameters}

\item{username}{(Optional if interactive) Username can be anything.
It's used to distinguish between users who share an API key}

\item{key}{(Optional if interactive) Azure API key}
}
\value{
Message
}
\description{
Setting up the Azure API info and storing it in the keyring
}
